requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-environ==0.9.0
future==0.18.2
html5lib==1.1
idna==3.4
kombu==5.2.4
lxml==4.9.1
oauthlib==3.2.1
oscrypto==1.3.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.5.0
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.3.4
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stripe==4.2.0
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
types-cryptography==3.3.23
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
